Add Windows MSI installer, fix x64 warnings.
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@10060 7a7537e8-13f0-0310-91df-b6672ffda945

+++ filter/interpret.c
@@ -1387,7 +1387,7 @@ setpagedevice(
   * Found the start of the dictionary, empty the stack to this point...
   */
 
-  st->num_objs = obj - st->objs;
+  st->num_objs = (int)(obj - st->objs);
 
  /*
   * Now pull /name and value pairs from the dictionary...
